+++ static/index.d.ts
@@ -1,0 +1,2 @@
+import type { Elysia } from 'elysia'
+export declare function static_middleware_():Promise<Elysia>